Rename image-block to image
