+++ packages/app/src/components-styled/cms/image.tsx
